Improve greedy evaluation.
Fixes the issue I noted in https://github.com/tkoolen/SimpleQP.jl/pull/37#pullrequestreview-132220452.

I opted to not try to fix `optimizearg`, but instead to simply get rid of it, and to allow `@expression` to return a *value* (instead of always returning a LazyExpression), if none of the arguments of the LazyExpression are parameters or other LazyExpressions.

To support this, I added a `convert`-to-`WrappedExpression` method that takes a value, and used it in the Objective and Constraint constructors.

This also fixes the broken allocation test case.